fix(ci): resolve all CI failures on PR #636
Backend: add pub mod orchestrator, fix api_error semicolons, fix models duplicate imports,
fix idempotency serde derives and IdempotencyKeyRequest

Frontend: add Image import, fix conditional useCallback, add a11y keyboard handler,
fix analytics formatter type, friends page isOnline, dashboard profile MatchWithPlayers type,
export AnyMatchWithPlayers type

Server: add module declarations for hpp and @sentry/node, fix pool types in database.service,
fix Prisma/Kafka type casts in audit.service and kafka/event-store

Contracts: add Val import in contract-utils, add Vec import + u32 decimals in contract-standards,
fix Vec::<Address> in token-manager, add missing event types in arenax-events,
fix duplicate functions and String::from_str in tournament-manager

Package-lock: add missing deps (react-window, kafkajs, react-dropzone, playwright, etc.)

+++ contracts/contract-standards/src/lib.rs
@@ -1,6 +1,6 @@
 #![no_std]
 
-use soroban_sdk::{contracttype, Address, Env, Map};
+use soroban_sdk::{contracttype, Address, Env, Map, Vec};
 
 // ---------------------------------------------------------------------------
 // Standardized Contract Interface Traits
@@ -129,14 +129,14 @@ macro_rules! impl_pausable {
 pub struct TokenMetadata {
     pub name: soroban_sdk::String,
     pub symbol: soroban_sdk::String,
-    pub decimals: u8,
+    pub decimals: u32,
 }
 
 /// Standard Token Interface (Soroban compatible)
 pub trait Token {
     fn name(env: &Env) -> soroban_sdk::String;
     fn symbol(env: &Env) -> soroban_sdk::String;
-    fn decimals(env: &Env) -> u8;
+    fn decimals(env: &Env) -> u32;
     fn total_supply(env: &Env) -> i128;
     fn balance(env: &Env, id: Address) -> i128;
     fn transfer(env: &Env, from: Address, to: Address, amount: i128);
